Move to a new versioning scheme
Releases are now numbered sequentially.  The master branch always has
the version of the next release.  When the new release is ready, it is
marked with an annotated tag and the version is incremented to start
development of the new release.

The version file is augmented by the file version.git, which holds the
result of "git describe" for pre-release versions.  This is obtained
either from Hydra via release.nix or from a local checkout of the
repo.

The onieInstaller derivation is only built for proper releases, based
on gitTag passed by Hydra.

The input in spec.json must include "deepClone" to prevent Hydra from
deleting .git after cloning the repo, which would make the tags
unavailable to "git describe".