release.nix: dont' build any installers by default
The idea is to enable buildInstaller only on release branches. There
will be one ONIE installer per release (for the initial release) and
one standalone Installer per update to the release.

Explicitly disable buildInstaller in spec.json for the master branch.

Rename the "packet-broker" jobset to "development". Jobsets for
releases will be named "release-<n>".  The Hydra project can have
"packet-broker" in its name, we don't have to repeat it here.