Add python3, python3-apt to ONIE image
These packages are needed to make Ansible work right after
installation (provided one also adds a suitable user account).

Unfortunately, we also have to move to a newer Debian release. This is
because the debootstrap command used by the onie-debian-nix-installer
can only create an image for the current Debian version at the time
the mk-profile.sh command is run.