Add support for the Tofino model
The packet broker can now be installed on a system without Tofino
ASIC, e.g. a VM. In that case, the program is run on the Tofino model
emulator. This can be useful for testing.

To use this feature, create /etc/machine.conf with

onie_machine=model

then execute "make install".