Replace bf-sde-nixpkgs by a version based on open-p4studio
open-p4studio is the open-source version of SDE 9.13.4. It excludes
some parts that are currently still encumbered by IP. The
open-p4studio Nix package provides the full SDE when built in "ASIC
mode", which requires access to the closed-source SDE and BSP.

Currently, this version of bf-sde-nixpkgs is kept on a separate branch
"open-p4studio".

The version introduced here contains a new version of the Asterfusion
BSP which supports their Tofino2-based model X732Q-T.