Update bf-sde-nixpkgs to ead7350
This fixes a number of issues with the Asterfusion BSP. Also change
the ONIE machine name of the Asterfusion X732Q-T to
asterfusion_x732q_t.

The change from the Intel SDE to open-p4studio also introduced a
change in the naming of some of the SDE packages. In particular,
bf-target is now called tofino-target (the old name was actually
incorrect but kept for historical reasons which are now obsolete).